backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
replicate
requests
moviepy
python-dotenv
httpx
databricks-sdk
mlflow
pydantic
pillow
numpy
